requirements.txt
torch>=2.0.7,<3.0
torchvision>=0.15,<1.0
transformers>=5.0.0rc1,<6.0
accelerate>=0.20,<1.0
numpy>=1.24,<3.0
rawpy>=0.25,<1.0
opencv-python>=4.8,<5.0
Pillow>=10.0,<11.0
PyYAML>=6.0,<7.0
huggingface_hub[cli]>=1.2.1,<2.0
